Fix: throw error when trying inexistant logs folder

+++ scripts/clear_logs.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 
 
@@ -5,9 +6,10 @@ def clear_logs():
     """
     Clears all logs
     """
-    shutil.rmtree('logs')
+    if os.path.exists("logs"):
+        shutil.rmtree("logs")
     print("Logs cleared")
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     clear_logs()
